+++ test/integration/react-streaming-and-server-components/test/rsc.js
@@ -3,12 +3,14 @@ import webdriver from 'next-webdriver'
 import cheerio from 'cheerio'
 import { renderViaHTTP, check } from 'next-test-utils'
 
+import { distDir } from './utils'
+
 function getNodeBySelector(html, selector) {
   const $ = cheerio.load(html)
   return $(selector)
 }
 
-export default function (context, { runtime }) {
+export default function (context, { runtime, env }) {
   it('should render server components correctly', async () => {
     const homeHTML = await renderViaHTTP(context.appPort, '/', null, {
       headers: {
@@ -69,6 +71,22 @@ export default function (context, { runtime }) {
       )
 
       expect(imageTag.attr('src')).toContain('data:image')
+    })
+  }
+
+  // For prod build, the directory contains the build ID so it's not deterministic.
+  // Only enable it for dev for now.
+  if (env === 'dev') {
+    it('should not bundle external imports into client builds for RSC', async () => {
+      const html = await renderViaHTTP(context.appPort, '/external-imports')
+      expect(html).toContain('date:')
+
+      const distServerDir = join(distDir, 'static', 'chunks', 'pages')
+      const bundle = fs
+        .readFileSync(join(distServerDir, 'external-imports.js'))
+        .toString()
+
+      expect(bundle).not.toContain('moment')
     })
   }
